Refactor and move package manager extensions to their own file
